feat: add support for configurable track scale and decimal lane lengths. Added ft indicator on lane length selector

+++ server/src/main/java/com/antigravity/models/Lane.java
@@ -9,13 +9,13 @@ public class Lane extends Model {
 
   private final String background_color;
   private final String foreground_color;
-  private final int length;
+  private final double length;
 
   @JsonCreator
   public Lane(
       @JsonProperty("background_color") String background_color,
       @JsonProperty("foreground_color") String foreground_color,
-      @JsonProperty("length") int length,
+      @JsonProperty("length") double length,
       @JsonProperty("entity_id") String entityId,
       @JsonProperty("_id") String id) {
     super(id, entityId);
@@ -24,7 +24,7 @@ public class Lane extends Model {
     this.length = length;
   }
 
-  public Lane(String background_color, String foreground_color, int length) {
+  public Lane(String background_color, String foreground_color, double length) {
     this(background_color, foreground_color, length, null, null);
   }
 
@@ -36,7 +36,7 @@ public class Lane extends Model {
     return foreground_color;
   }
 
-  public int getLength() {
+  public double getLength() {
     return length;
   }
 }
